hotkey: Re-factor into a module, per-arch files
This helps avoid some pre-commit/mypy carping.
